verifact_system/requirements.txt
# Core dependencies - Python 3.8 compatible
fastapi==0.88.0
uvicorn==0.20.0
streamlit==1.28.1
pandas==1.5.3
numpy==1.24.3
python-dotenv==1.0.0

# Database - versioni compatibili
psycopg2-binary==2.9.7
redis==4.6.0
sqlalchemy==1.4.48

# Alternative MongoDB support
pymongo==4.5.0

# ML/AI Stack - versioni Python 3.8
torch==1.13.1
transformers==4.21.3
sentence-transformers==2.2.2
datasets==2.14.6
scikit-learn==1.3.0

# NLP Processing
spacy==3.6.1
nltk==3.8.1
langdetect==1.0.9
clean-text==0.6.0
ftfy==6.1.1

# Vector Database - versioni compatibili
chromadb==0.3.29

# Web scraping
requests==2.31.0
aiohttp==3.8.5
beautifulsoup4==4.12.2

# Data processing
tqdm==4.65.0
pyyaml==6.0
openpyxl==3.1.2

# Visualization
plotly==5.15.0
matplotlib==3.7.2
seaborn==0.12.2

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1

# Development tools
black==23.7.0
flake8==6.0.0